front end: fixes from a read-only audit, including one that made it dead code
An audit pass over the new module found real defects. Nothing here has met a
compiler, so this is the closest available substitute, and it earned its keep
on the first finding alone.

THE FRONT END NEVER ACTIVATED. State defaults to Inactive; Initialize assigned
Inactive on the SUPPRESSED branch and assigned nothing on the enabled one; and
OnWorldBeginPlay opens with `if (State == Inactive) return;`. So the log said
"VoxelFrontEnd: active", the early-out fired anyway, IsTickable() was false
forever, and no menu would ever have appeared -- with nothing in the log
admitting it. "Enabled" and "not started yet" now have distinct states, and
OnWorldBeginPlay tests the POLICY rather than a state that has already been
mis-assigned once.

Three more that would have shipped:

  * MipGenSettings is inside WITH_EDITORONLY_DATA. Unguarded, it compiles in
    the editor target and FAILS in Game and packaged client -- which is exactly
    what a ClientOnly module exists to serve.
  * FGeometry::ToPaintGeometry() returns by value and
    GetAccumulatedRenderTransform() returns a reference INTO it. Binding that
    reference through the call chain does not extend the temporary's lifetime,
    so every hourglass vertex was being transformed by freed memory. Compiles
    clean, draws garbage.
  * FNavigationMetaData was attached to the button column. Slate reads it off
    the widget that HAS focus, not off an ancestor, so the end-of-list wrap
    would never have been consulted and pressing Up on the first button would
    have done nothing -- which is indistinguishable from a broken menu. It now
    lives on each button's focusable inner SButton.

And four hardening changes:

  * A declared out-of-line destructor for the subsystem. ReadyProbe is a
    TUniquePtr to a type the generated FVTableHelper constructor only sees as
    incomplete, which deletes through it without running ~FVoxelWorldReadyProbe.
  * The async decode continuation captured raw `this` and an index, guarded
    only by a null check -- which passes after Shutdown()+Get() has built a
    different instance at a different address, and says nothing about
    RescanBackgrounds having reset Entries underneath it. Now guarded by
    instance identity, a generation counter and an index check.
  * IImageWrapper::GetRaw takes TArray64; the TArray convenience overload has
    been on a deprecation track. Using the 64-bit form is both future-proof and
    one fewer copy.
  * The decoded backgrounds are released at hand-off. The front end is the only
    thing that draws them and roughly 48 MB is not an amount to hold for the
    rest of a session out of politeness.

backlog.md §11 gains the two API questions the audit could not settle without
an engine -- MakeLines' point-array type, and whether DoesSupportWorldType's
EWorldType form has been superseded. If the second one is wrong it is wrong in
the pre-existing VoxelWorldSubsystem too, which is worth knowing either way.

+++ ue-project/Source/VoxelEarthUI/VoxelWorldReadyProbe.cpp
@@ -112,7 +112,7 @@ void FVoxelWorldReadyProbe::Poll(const UVoxelWorldSubsystem& World)
 		Pending += Progress.LevelPendingCount[Level];
 		Jobs += Progress.LevelJobsInFlight[Level];
 
-		const float Weight = Level < UE_ARRAY_COUNT(VoxelReadyProbeDetail::kRingWeights)
+		const float Weight = Level < int32(UE_ARRAY_COUNT(VoxelReadyProbeDetail::kRingWeights))
 		                         ? VoxelReadyProbeDetail::kRingWeights[Level]
 		                         : 0.1f;
 		const int32 Loaded = Progress.LevelLoadedCount[Level];
